Unify api step and result elements controllers implementation and types [SCI-9907]
