+++ app/serializers/api/v2/result_table_serializer.rb
@@ -2,21 +2,8 @@
 
 module Api
   module V2
-    class ResultTableSerializer < ActiveModel::Serializer
+    class ResultTableSerializer < Api::V1::ResultTableSerializer
       type :tables
-      attributes :table_id, :table_contents, :table_metadata
-
-      def table_id
-        object.table&.id
-      end
-
-      def table_contents
-        object.table&.contents&.force_encoding(Encoding::UTF_8)
-      end
-
-      def table_metadata
-        object.table&.metadata
-      end
     end
   end
 end
