+++ app/serializers/api/v2/step_text_serializer.rb
@@ -1,0 +1,9 @@
+# frozen_string_literal: true
+
+module Api
+  module V2
+    class StepTextSerializer < Api::V1::StepTextSerializer
+      type :texts
+    end
+  end
+end
