+++ spec/requests/api/v2/result_elements/tables_controller_spec.rb
@@ -4,7 +4,7 @@
 
 require 'rails_helper'
 
-RSpec.describe 'Api::V2::ResultTablesController', type: :request do
+RSpec.describe 'Api::V2::ResultElements::TablesController', type: :request do
   let(:user) { create(:user) }
   let(:team) { create(:team, created_by: user) }
   let(:project) { create(:project, team: team, created_by: user) }
@@ -15,7 +15,7 @@ RSpec.describe 'Api::V2::ResultTablesController', type: :request do
   let(:valid_headers) { { Authorization: "Bearer #{generate_token(user.id)}", 'Content-Type': 'application/json' } }
 
   let(:api_path) do
-    api_v2_team_project_experiment_task_result_result_tables_path(
+    api_v2_team_project_experiment_task_result_tables_path(
       team_id: team.id,
       project_id: project.id,
       experiment_id: experiment.id,
@@ -46,7 +46,7 @@ RSpec.describe 'Api::V2::ResultTablesController', type: :request do
 
     context 'when result is not found' do
       it 'renders 404' do
-        get api_v2_team_project_experiment_task_result_result_tables_path(
+        get api_v2_team_project_experiment_task_result_tables_path(
           team_id: team.id,
           project_id: project.id,
           experiment_id: experiment.id,
@@ -65,7 +65,7 @@ RSpec.describe 'Api::V2::ResultTablesController', type: :request do
     context 'when has valid params' do
       it 'renders 200' do
         hash_body = nil
-        get api_v2_team_project_experiment_task_result_result_table_path(
+        get api_v2_team_project_experiment_task_result_table_path(
           team_id: team.id,
           project_id: project.id,
           experiment_id: experiment.id,
@@ -134,7 +134,7 @@ RSpec.describe 'Api::V2::ResultTablesController', type: :request do
       let(:request_body) do
         {
           data: {
-            type: 'result_tables',
+            type: 'tables',
             attributes: {}
           }
         }
@@ -152,7 +152,7 @@ RSpec.describe 'Api::V2::ResultTablesController', type: :request do
     let(:result_table) { create(:result_table, result: result) }
     let(:result_table_archived) { create(:result_table, result: result_archived) }
     let(:action) do
-      patch(api_v2_team_project_experiment_task_result_result_table_path(
+      patch(api_v2_team_project_experiment_task_result_table_path(
         team_id: team.id,
         project_id: project.id,
         experiment_id: experiment.id,
@@ -163,7 +163,7 @@ RSpec.describe 'Api::V2::ResultTablesController', type: :request do
     end
 
     let(:action_archived) do
-      patch(api_v2_team_project_experiment_task_result_result_table_path(
+      patch(api_v2_team_project_experiment_task_result_table_path(
         team_id: team.id,
         project_id: project.id,
         experiment_id: experiment.id,
@@ -211,7 +211,7 @@ RSpec.describe 'Api::V2::ResultTablesController', type: :request do
       let(:request_body) do
         {
           data: {
-            type: 'result_tables',
+            type: 'tables',
             attributes: {}
           }
         }
@@ -228,7 +228,7 @@ RSpec.describe 'Api::V2::ResultTablesController', type: :request do
       let(:request_body) do
         {
           data: {
-            type: 'result_tables',
+            type: 'tables',
             attributes: {
               name: 'Result table',
               contents: '{"data": [["group/time", "1 dpi", "6 dpi", "", ""], ["PVYNTN", "1", "1", "", ""]]}'
@@ -249,7 +249,7 @@ RSpec.describe 'Api::V2::ResultTablesController', type: :request do
     let(:result_table) { create(:result_table, result: result) }
     let(:result_table_archived) { create(:result_table, result: result_archived) }
     let(:delete_action) do
-      delete(api_v2_team_project_experiment_task_result_result_table_path(
+      delete(api_v2_team_project_experiment_task_result_table_path(
         team_id: team.id,
         project_id: project.id,
         experiment_id: experiment.id,
@@ -260,7 +260,7 @@ RSpec.describe 'Api::V2::ResultTablesController', type: :request do
     end
 
     let(:delete_action_archived) do
-      delete(api_v2_team_project_experiment_task_result_result_table_path(
+      delete(api_v2_team_project_experiment_task_result_table_path(
         team_id: team.id,
         project_id: project.id,
         experiment_id: experiment.id,
