+++ spec/requests/api/v2/step_elements/checklists_controller_spec.rb
@@ -2,7 +2,7 @@
 
 require 'rails_helper'
 
-RSpec.describe 'Api::V2::ChecklistsController', type: :request do
+RSpec.describe 'Api::V2::StepElements::ChecklistsController', type: :request do
   before :all do
     @user = create(:user)
     @team = create(:team, created_by: @user)
